requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.6
fastapi-cli==0.0.6
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.1.3
oauthlib==3.2.2
openai==1.55.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
XlsxWriter==3.2.0
langfuse~=2.58.0